pyproject.toml
[tool.poetry]
name = "BILIBILI Record Monitor"
version = "1.1.1"
description = "基于 OneDrive Client for Linux 和 watchdog 的BILIBILI录播文件监视器"
authors = ["Nya-WSL"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9,<3.14"
watchdog = "^6.0.0"
pyyaml = "^6.0.2"
wxpusher = "^2.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
